Add method to auto-expire assignments in AssignmentManager
A new method `CheckAndExpireAssignments` has been added to the
`AssignmentManager` class. This method retrieves all assignments
and checks if any have expired based on their `CompletionTime`
and `CompletionStatus`. Expired assignments are updated to
`Expired` status, and observers are notified of these changes
as well as the overall update to the assignment list. This
ensures assignments are kept up-to-date automatically and
observers can react to these changes appropriately.